KNIFE-394: Fix eventmachine dependency
We formerly used =1.0.0.beta.3 to ensure specific functionality,
but 1.0.0 has been out for some time. This will accept a great range
of releases as a match, while being "pessimistic" about future releases.

If the release cadence was faster, I would consider risking ~> 1.0, but this
seems appropriate, provided we pay better attention in the future.